Fixed signed/unsigned integer comparison issue.

+++ moses/FF/GlobalLexicalModelUnlimited.cpp
@@ -111,7 +111,7 @@ void GlobalLexicalModelUnlimited::Evaluate(const Hypothesis& cur_hypo, ScoreComp
   const Sentence& input = *(m_local->input);
   const TargetPhrase& targetPhrase = cur_hypo.GetCurrTargetPhrase();
 
-  for(int targetIndex = 0; targetIndex < targetPhrase.GetSize(); targetIndex++ ) {
+  for(size_t targetIndex = 0; targetIndex < targetPhrase.GetSize(); targetIndex++ ) {
     StringPiece targetString = targetPhrase.GetWord(targetIndex).GetString(0); // TODO: change for other factors
 
     if (m_ignorePunctuation) {
@@ -132,8 +132,9 @@ void GlobalLexicalModelUnlimited::Evaluate(const Hypothesis& cur_hypo, ScoreComp
     }
 
     boost::unordered_set<uint64_t> alreadyScored;
-    for(int sourceIndex = 0; sourceIndex < input.GetSize(); sourceIndex++ ) {
-      const StringPiece sourceString = input.GetWord(sourceIndex).GetString(0); // TODO: change for other factors
+    for(size_t sourceIndex = 0; sourceIndex < input.GetSize(); sourceIndex++ ) {
+      const StringPiece sourceString = input.GetWord(sourceIndex).GetString(0); 
+      // TODO: change for other factors
 
       if (m_ignorePunctuation) {
         // check if first char is punctuation
